Made shallow cloning not so shallow.
As the original object was linked to the clone through the prototype-
chain, changes in the original would propagate to its clones.

We are not creating a real blank object using the original object's-
constructor's prototype passed to Object.create() and manually assigning
all enumerable and non-enumerable values to the cloned object using-
Object.defineProperty(). This ensures all properties dynamically created
by the original are copied with proper access control.